pyproject.toml
[tool.poetry]
name = "project_name"
version = "0.0.1"

[tool.poetry.dependencies]
python = "^3.8"
Django = "~3.0.0"
django-split-settings = "^1.0.1"
django-environ = "^0.4.5"
django-filter = "^2.4.0"
django-mysql = "^3.9.0"
mysqlclient = "pytest"
requests = "^2.24.0"
gunicorn = "^20.0.4"
celery = "~4.4.0"
redis = "^3.5.3"
django-extensions = "^3.0.9"
drf-yasg = "^1.20.0"
django-cors-headers = "^3.5.0"
djangorestframework-simplejwt = "^4.5.0"
djoser = "^2.1.0"
django-celery-beat = "^2.1.0"
djangorestframework = "~=3.11.0"
django-redis = "^4.12.1"

[tool.poetry.dev-dependencies]
factory-boy = "^3.1.0"
flake8 = "^3.8.4"
autoflake = "^1.4"
bpython = "^0.20"
importmagic = "^0.1.7"
ipdb = "^0.13.4"
ipython = "^7.19.0"
pyls-black = "^0.4.6"
pyls-isort = "^0.2.0"
pytest = "^6.1.2"
pytest-mock = "^3.3.1"
pytest-cov = "^2.10.1"
pytest-django = "^4.1.0"
pytest-env = "^0.6.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py36', 'py37', 'py38']
include = '\.pyi?$'
exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''
